Create CDS
It is a kind of a credit derivative, and is a derivative derivative of the financial derivative. In the event that the target company collapses, and the payment of financial claims and corporate bonds cannot be paid, the purchase of the CDs will receive the payment corresponding to interest rate and principal.